oops, moved DriftState to model
