+++ morpheus-plugin-api/src/main/java/com/morpheusdata/core/providers/NetworkProvider.java
@@ -16,12 +16,11 @@
 
 package com.morpheusdata.core.providers;
 
-import com.morpheusdata.core.DriftState;
+import com.morpheusdata.model.DriftState;
 import com.morpheusdata.model.*;
 import com.morpheusdata.response.ServiceResponse;
 import com.morpheusdata.views.HTMLResponse;
 
-import java.security.Provider;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
